+++ src/eval.h
@@ -31,9 +31,8 @@ void evalQuit(void);
 Score evaluate(const Pos *pos); // Returns score in CP.
 
 // Texel Tuning functions
-void evaluateCoefficients(const Pos *pos, double *coefficients);
+void evaluateCoefficients(const Pos *pos, float *coefficients);
 void evaluateOutputCode(const char *path, const int *weights);
-
 
 void evalClear(void); // Clear all saved data (called when we receive 'ucinewgame', for example).
